+++ WooCommerce/WooCommerceTests/Testing/Product+Helpers.swift
@@ -1,0 +1,17 @@
+import WooCommerce
+import Yosemite
+
+extension Product {
+    func toOrderDetailsProduct() -> OrderDetailsProduct {
+        OrderDetailsProduct(siteID: siteID,
+                            productID: productID,
+                            name: name,
+                            productTypeKey: productTypeKey,
+                            sku: sku,
+                            price: price,
+                            virtual: virtual,
+                            stockQuantity: stockQuantity,
+                            imageURL: imageURL,
+                            addOns: addOns)
+    }
+}
